+++ src/aggregator.ts
@@ -31,7 +31,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
 import { DataFrame, FieldType, Labels, MutableDataFrame } from '@grafana/data';
+import { SlidingAccumulator } from 'sliding';
 import { AggregationSpec, KeyValue } from 'types';
+const TDigest = require('tdigest').TDigest;
 
 const statProducers: KeyValue = {
   avg: (acc: Accumulator) => acc.getAverage(),
@@ -41,8 +43,12 @@ const statProducers: KeyValue = {
   sum: (acc: Accumulator) => acc.getSum(),
   count: (acc: Accumulator) => acc.getCount(),
   variance: (acc: Accumulator) => acc.getVariance(),
+  percentile: (acc: Accumulator) => acc.getPercentile(),
 };
 
+/**
+ * Statistics accumulator for calculating average, variance, standard deviation and percentiles
+ */
 export class Accumulator {
   sum = 0;
   count = 0;
@@ -50,8 +56,23 @@ export class Accumulator {
   min: number = Number.MAX_VALUE;
   vAcc = 0;
   avg = 0;
+  wantDigest = false;
+  digest: any;
+  percentile? = 0.0;
 
-  aaddDataPoint(value: number) {
+  constructor(wantPercentile: boolean, percentile?: number) {
+    this.wantDigest = wantPercentile;
+    if (wantPercentile) {
+      this.digest = new TDigest();
+    }
+    this.percentile = percentile;
+  }
+
+  /**
+   * Adds a new sample to the accumulator
+   * @param value
+   */
+  addDataPoint(value: number) {
     this.sum += value;
     this.count++;
     this.max = Math.max(this.max, value);
@@ -59,6 +80,9 @@ export class Accumulator {
     const avg = this.sum / this.count;
     this.vAcc += (value - this.avg) * (value - avg);
     this.avg = avg;
+    if (this.wantDigest) {
+      this.digest.push(value);
+    }
   }
 
   getAverage(): number {
@@ -88,20 +112,44 @@ export class Accumulator {
   getStandardDeviation(): number {
     return Math.sqrt(this.getVariance());
   }
+
+  getPercentile(): number {
+    this.digest.compress();
+    return this.digest.percentile(this.percentile! / 100);
+  }
 }
 
+/**
+ * Holds the accumulators for a specific point in time.
+ */
 export class Bucket {
   accumulators: Map<number, Accumulator> = new Map();
+  wantPercentile = false;
+  percentile: number | undefined = 0.0;
 
+  constructor(wantPercentile: boolean, percentile: number) {
+    this.percentile = percentile;
+    this.wantPercentile = wantPercentile;
+  }
+
+  /**
+   * Adds a new sample
+   * @param timestamp
+   * @param value
+   */
   addDataPoint(timestamp: number, value: number) {
     let accumulator = this.accumulators.get(timestamp);
     if (!accumulator) {
-      accumulator = new Accumulator();
+      accumulator = new Accumulator(this.wantPercentile, this.percentile);
       this.accumulators.set(timestamp, accumulator);
     }
-    accumulator.aaddDataPoint(value);
+    accumulator.addDataPoint(value);
   }
 
+  /**
+   * Returns the accumulators for each point in time, sorted by time.
+   * @returns
+   */
   getResults(): Map<number, Accumulator> {
     // If datapoints are missing in some time series, it's possible that some
     // values are inserted out of order. To prevent strange graph artifacts,
@@ -110,9 +158,26 @@ export class Bucket {
   }
 }
 
+/**
+ * Statistics accumulator for an entire composite time-series
+ */
 export class Stats {
+  wantPercentile = false;
+  percentile = 0.0;
+  constructor(aggregation: AggregationSpec) {
+    if (aggregation.type === 'percentile') {
+      this.wantPercentile = true;
+      this.percentile = aggregation.parameter ? aggregation.parameter : 0.0;
+    }
+  }
   buckets: Map<string, Bucket> = new Map();
 
+  /**
+   * Adds a sample and updates the accumulators
+   * @param timestamps
+   * @param values
+   * @param properties
+   */
   add(timestamps: number[], values: number[], properties: Map<string, string>) {
     const key = JSON.stringify(Array.from(properties?.entries() || {}));
     for (const idx in timestamps) {
@@ -120,14 +185,26 @@ export class Stats {
       const value = values[idx];
       let slot = this.buckets.get(key);
       if (!slot) {
-        slot = new Bucket();
+        slot = new Bucket(this.wantPercentile, this.percentile);
         this.buckets.set(key, slot);
       }
       slot.addDataPoint(ts, value);
     }
   }
 
-  toFrames(refId: string, aggregation: AggregationSpec): DataFrame[] {
+  /**
+   *
+   * @param refId Converts the computed statistical key figures. If slidingWindowFactoey is
+   * non-null, the specified sliding window is applied to the resulting time series.
+   * @param aggregation
+   * @param slidingWindowFactory
+   * @returns
+   */
+  toFrames(
+    refId: string,
+    aggregation: AggregationSpec,
+    slidingWindowFactory: (() => SlidingAccumulator) | null
+  ): DataFrame[] {
     const produce = statProducers[aggregation.type];
     if (!produce) {
       throw 'Internal error: Producer ' + aggregation.type + ' not found';
@@ -135,6 +212,9 @@ export class Stats {
     const frames: MutableDataFrame[] = [];
     let statKey = '<undefined>';
     for (const [key, bucket] of this.buckets) {
+      const slidingWindow = slidingWindowFactory
+        ? slidingWindowFactory()
+        : null;
       const labels: Labels = {};
       if (key) {
         const properties = new Map<string, string>(JSON.parse(key));
@@ -155,7 +235,11 @@ export class Stats {
         ],
       });
       for (const [timestamp, data] of bucket.getResults()) {
-        frame.add({ Time: timestamp, Value: produce(data) });
+        const value = produce(data);
+        const point = slidingWindow
+          ? slidingWindow.pushAndGet(timestamp, value)
+          : value;
+        frame.add({ Time: timestamp, Value: point });
       }
       frames.push(frame);
     }